// Avocat_Angular-master/src/app/models/courrier.ts
export class Courrier {

  numeroCourrier: string;
  dateCourrier: string;
  typeCourrier: string;
  reponse: string;
  parVoie: string;
  destinataire: string;
  degre: string;
  reponseExigee: boolean;
  sujet: string;
  expediteur: string;
  lien: string;
  numOrigine: number;
  dateOrigine: string;
  numeroAO_CS: number;
  finDepot: string;
  consultation: boolean;
  AO_International: boolean;
  etatAvancement: string;
  type: string;
  pdfContent: Uint8Array;
  thumbnailPath: string;

}